requirements.txt
python-telegram-bot
pandas
firebase_admin
pytest